co-locate grpc methods
