--- packages/insomnia/src/network/grpc/__schemas__/grpc-ipc-message-params-schema.ts
@@ -1,10 +1,0 @@
-import type { Schema } from '@develohpanda/fluent-builder';
-
-import { GrpcIpcMessageParams } from '..';
-
-export const grpcIpcMessageParamsSchema: Schema<GrpcIpcMessageParams> = {
-  requestId: () => 'gr',
-  body: () => ({
-    text: '{}',
-  }),
-};
