--- packages/insomnia/src/network/grpc/__schemas__/grpc-ipc-request-params-schema.ts
@@ -1,8 +1,0 @@
-import type { Schema } from '@develohpanda/fluent-builder';
-
-import { GrpcIpcRequestParams } from '..';
-
-export const grpcIpcRequestParamsSchema: Schema<GrpcIpcRequestParams> = {
-  // @ts-expect-error -- TSCONVERSION
-  request: () => ({}),
-};
